+++ backend/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+authors = ["Sridhar Ratnakumar <srid@srid.ca>"]
+edition = "2021"
+# If you change the name here, you must also do it in flake.nix (and run `cargo generate-lockfile` afterwards)
+name = "backend"
+version = "0.1.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+axum = { version = "0.6.19", features = ["json", "tokio"] }
+serde = { version = "1.0.175", features = ["derive"] }
+serde_json = "1.0.103"
+tokio = { version = "1.29.1", features = ["full"] }
