+++ backend/src/main.rs
@@ -1,0 +1,17 @@
+use axum::{routing::get, Router};
+use std::net::SocketAddr;
+
+#[tokio::main]
+async fn main() {
+    let app = Router::new().route("/", get(root));
+    let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
+    println!("Launching http://localhost:3000");
+    axum::Server::bind(&addr)
+        .serve(app.into_make_service())
+        .await
+        .unwrap();
+}
+
+async fn root() -> &'static str {
+    "Hello, World!"
+}
