+++ frontend/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+authors = ["Sridhar Ratnakumar <srid@srid.ca>"]
+edition = "2021"
+# If you change the name here, you must also do it in flake.nix (and run `cargo generate-lockfile` afterwards)
+name = "frontend"
+version = "0.1.0"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0.175", features = ["derive"] }
+serde_json = "1.0.103"
+wasm-bindgen = "0.2.84"
+# vanilla wasm app
+console_error_panic_hook = "0.1"
+web-sys = { version = "0.3", features = [
+    "Window",
+    "Document",
+    "HtmlElement",
+    "Node",
+    "Text",
+] }
