+++ frontend/Trunk.toml
@@ -1,0 +1,3 @@
+[[proxy]]
+rewrite = "/api"
+backend = "http://localhost:3000"
